Added giant invisible button!
This button allows the playback controls to be hidden when video is
tapped, just like in iOS! Kodi's normal method to hide playback controls
is to simply let the controls fade away after several seconds of
inactivity which is annoying if you just want those controls out of your
face. This is the last feature before the code cleanup, yay!